+++ src/components/animation.ts
@@ -1,7 +1,12 @@
 import * as types from './types';
 import * as api from './api';
+import { updateCarsBtns } from './common';
 
 const START_TRANSLATE = 65;
+const COEFF_M_TO_KM = 1000;
+const MIN_WINDOW_WIDTH = 500;
+const MAX_CAR_WIDTH = 130;
+const MIN_CAR_WIDTH = 100;
 
 export async function stopCar(id: number | null, car: HTMLElement) {
   if (id) {    
@@ -13,39 +18,27 @@ export async function stopCar(id: number | null, car: HTMLElement) {
 export async function driveCar(id: number | null, car: HTMLElement) {
   if (id) {    
     const windowWidth = window.innerWidth;
-    const carWidth = windowWidth > 500 ? 130 : 100;
-
+    const carWidth = windowWidth > MIN_WINDOW_WIDTH ? MAX_CAR_WIDTH : MIN_CAR_WIDTH;
     const startResponce = await api.startEngine(id);
-    let fullAnimationTime = (startResponce.distance / 1000) / startResponce.velocity;
+    let fullAnimationTime = (startResponce.distance / COEFF_M_TO_KM) / startResponce.velocity;
+    const driveResponce = await api.driveEngine(id);
     car.style.transform = `translateX(${Number(window.innerWidth) - carWidth}px)`;
     car.style.transitionDuration = `${Math.round(fullAnimationTime)}s`;
-
-    const driveResponce = await api.driveEngine(id);
-    //if (!driveResponce) stopBrokenCar(id, car);
     console.log(driveResponce);
+    if (!driveResponce) stopBrokenCar(id, car);
   }
 }
 
-// function stopBrokenCar(id: number, car: HTMLElement) {
-//   const duration = car;
-//   console.log('duration:::', duration);
-//   car.style.transform = `translateX(${duration}px)`;
-//   car.style.transitionDuration = `20s`;
-// }
+function stopBrokenCar(id: number, car: HTMLElement) {
+  const duration = car.style.transform;
+  console.log('duration:::', duration);
+  car.style.transform = duration;
+  car.style.transitionDuration = `2s`;
+}
 
 function returnCarToGarage(car: HTMLElement) {
   car.style.transform = `translateX(${START_TRANSLATE}px)`;
   car.style.transitionDuration = `${0}s`;
-}
-
-export function updateCarsBtns(firstBtn: HTMLElement, secondBtn: HTMLElement) {
-  if ((firstBtn as HTMLButtonElement).disabled === true) {
-    (firstBtn as HTMLButtonElement).disabled = false;
-    (secondBtn as HTMLButtonElement).disabled = true;
-  } else if ((firstBtn as HTMLButtonElement).disabled === false) {
-    (firstBtn as HTMLButtonElement).disabled = true;
-    (secondBtn as HTMLButtonElement).disabled = false;
-  }
 }
 
 export function controlRace(type: string) {
