+++ src/style.scss
@@ -192,6 +192,8 @@ h2, h3 {
   transform: translateX(65px);
 }
 
+
+
 .buttons-row {
   display: flex;
   align-items: center;
